+++ jgrapht-core/src/main/java/org/jgrapht/traverse/RandomWalkIterator.java
@@ -1,5 +1,5 @@
 /*
- * (C) Copyright 2016-2020, by Assaf Mizrachi and Contributors.
+ * (C) Copyright 2020-2020, by Dimitrios Michail and Contributors.
  *
  * JGraphT : a free Java graph-theory library
  *
@@ -17,247 +17,150 @@
  */
 package org.jgrapht.traverse;
 
-import org.jgrapht.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+import java.util.NoSuchElementException;
+import java.util.Objects;
+import java.util.Random;
+import java.util.stream.Collectors;
 
-import java.util.*;
+import org.jgrapht.Graph;
+import org.jgrapht.Graphs;
 
 /**
- * A random walk iterator for a directed or undirected graph.
+ * A random walk iterator.
  * 
- * <p>
- * At each step the iterator selects a random (uniformly distributed) edge out of the current vertex
- * and follows it to the next vertex. In case of directed graphs the outgoing edge set is used. See
- * <a href="https://en.wikipedia.org/wiki/Random_walk#Random_walk_on_graphs">wikipedia</a> for more
- * details.
+ * "Given a graph and a starting point, we select a neighbor of it at random, and move to this
+ * neighbor; then we select a neighbor of this point at random, and move to it etc. The (random)
+ * sequence of points selected this way is a random walk on the graph." This very simple definition,
+ * together with a comprehensive survey can be found at: "Lovász, L. (1993). Random walks on graphs:
+ * A survey. Combinatorics, Paul erdos is eighty, 2(1), 1-46."
  * 
- * <p>
- * In case a weighted walk is desired, edges are selected with probability respective to its weight
- * (out of the total weight of the edges). The walk can be bounded by number of steps (default
- * {@code Long#MAX_VALUE} . When the bound is reached the iterator is considered exhausted. Calling
- * {@code next()} on exhausted iterator will throw {@code NoSuchElementException}.
+ * In its default variant the probability of selecting an outgoing edge is one over the (out) degree
+ * of the vertex. In case the user requests a weighted walk, then the probability of each edge is
+ * equal to its weight divided by the total weight of all outgoing edges. The walk can also be
+ * bounded by a maximum number of hops (edges traversed). The iterator returns
+ * {@link NoSuchElementException} when this bound is reached.
  * 
- * In case a sink (i.e. no edges) vertex is reached, any consecutive calls to {@code next()} will
- * throw {@code NoSuchElementException}.
- * 
- * <p>
- * For this iterator to work correctly the graph must not be modified during iteration. Currently
- * there are no means to ensure that, nor to fail-fast. The results of such modifications are
- * undefined.
- * 
- * @author Assaf Mizrachi
+ * @author Dimitrios Michail
  *
- * @param <V> vertex type
- * @param <E> edge type
+ * @param <V> the graph vertex type
+ * @param <E> the graph edge type
  */
 public class RandomWalkIterator<V, E>
-    extends
-    AbstractGraphIterator<V, E>
+    implements
+    Iterator<V>
 {
-    private V currentVertex;
-    private final boolean isWeighted;
-    private boolean sinkReached;
-    private long maxSteps;
-    private Random random;
+    private final Random rng;
+    private final Graph<V, E> graph;
+    private final boolean weighted;
+    private final Map<V, Double> outEdgesTotalWeight;
+    private final long maxHops;
+    private long hops;
+    private V nextVertex;
 
     /**
-     * Creates a new iterator for the specified graph. Iteration will start at arbitrary vertex.
-     * Walk is un-weighted and bounded by {@code Long#MAX_VALUE} steps.
-     *
-     * @param graph the graph to be iterated.
-     *
-     * @throws IllegalArgumentException if <code>graph==null</code> or does not contain
-     *         <code>startVertex</code>
+     * Create a new iterator
+     * 
+     * @param graph the graph
+     * @param vertex the starting vertex
      */
-    public RandomWalkIterator(Graph<V, E> graph)
+    public RandomWalkIterator(Graph<V, E> graph, V vertex)
     {
-        this(graph, null);
+        this(graph, vertex, Long.MAX_VALUE, false, new Random());
     }
 
     /**
-     * Creates a new iterator for the specified graph. Iteration will start at the specified start
-     * vertex. If the specified start vertex is <code>
-     * null</code>, Iteration will start at an arbitrary graph vertex. Walk is un-weighted and
-     * bounded by {@code Long#MAX_VALUE} steps.
-     *
-     * @param graph the graph to be iterated.
-     * @param startVertex the vertex iteration to be started.
-     *
-     * @throws IllegalArgumentException if <code>graph==null</code> or does not contain
-     *         <code>startVertex</code>
+     * Create a new iterator
+     * 
+     * @param graph the graph
+     * @param vertex the starting vertex
+     * @param maxHops maximum hops to perform during the walk
      */
-    public RandomWalkIterator(Graph<V, E> graph, V startVertex)
+    public RandomWalkIterator(Graph<V, E> graph, V vertex, long maxHops)
     {
-        this(graph, startVertex, true);
+        this(graph, vertex, maxHops, false, new Random());
     }
 
     /**
-     * Creates a new iterator for the specified graph. Iteration will start at the specified start
-     * vertex. If the specified start vertex is <code>
-     * null</code>, Iteration will start at an arbitrary graph vertex. Walk is bounded by
-     * {@code Long#MAX_VALUE} steps.
-     *
-     * @param graph the graph to be iterated.
-     * @param startVertex the vertex iteration to be started.
-     * @param isWeighted set to <code>true</code> if a weighted walk is desired.
-     *
-     * @throws IllegalArgumentException if <code>graph==null</code> or does not contain
-     *         <code>startVertex</code>
-     */
-    public RandomWalkIterator(Graph<V, E> graph, V startVertex, boolean isWeighted)
-    {
-        this(graph, startVertex, isWeighted, Long.MAX_VALUE);
-    }
-
-    /**
-     * Creates a new iterator for the specified graph. Iteration will start at the specified start
-     * vertex. If the specified start vertex is <code>
-     * null</code>, Iteration will start at an arbitrary graph vertex. Walk is bounded by the
-     * provided number steps.
-     *
-     * @param graph the graph to be iterated.
-     * @param startVertex the vertex iteration to be started.
-     * @param isWeighted set to <code>true</code> if a weighted walk is desired.
-     * @param maxSteps number of steps before walk is exhausted.
-     *
-     * @throws IllegalArgumentException if <code>graph==null</code> or does not contain
-     *         <code>startVertex</code>
-     */
-    public RandomWalkIterator(Graph<V, E> graph, V startVertex, boolean isWeighted, long maxSteps)
-    {
-        this(graph, startVertex, isWeighted, maxSteps, new Random());
-    }
-
-    /**
-     * Creates a new iterator for the specified graph. Iteration will start at the specified start
-     * vertex. If the specified start vertex is <code>
-     * null</code>, Iteration will start at an arbitrary graph vertex. Walk is bounded by the
-     * provided number steps.
-     *
-     * @param graph the graph to be iterated.
-     * @param startVertex the vertex iteration to be started.
-     * @param isWeighted set to <code>true</code> if a weighted walk is desired.
-     * @param maxSteps number of steps before walk is exhausted.
-     * @param rng the random number generator to use
-     *
-     * @throws IllegalArgumentException if <code>graph==null</code> or does not contain
-     *         <code>startVertex</code>
+     * Create a new iterator
+     * 
+     * @param graph the graph
+     * @param vertex the starting vertex
+     * @param maxHops maximum hops to perform during the walk
+     * @param weighted whether to perform a weighted random walk (compute probabilities based on the
+     *        edge weights)
+     * @param rng the random number generator
      */
     public RandomWalkIterator(
-        Graph<V, E> graph, V startVertex, boolean isWeighted, long maxSteps, Random rng)
+        Graph<V, E> graph, V vertex, long maxHops, boolean weighted, Random rng)
     {
-        super(graph);
-
-        // do not cross components.
-        this.crossComponentTraversal = false;
-        this.isWeighted = isWeighted;
-        this.maxSteps = maxSteps;
-
-        // select a random start vertex in case not provided.
-        if (startVertex == null) {
-            if (graph.vertexSet().size() > 0) {
-                currentVertex = graph.vertexSet().iterator().next();
-            }
-        } else if (graph.containsVertex(startVertex)) {
-            currentVertex = startVertex;
-        } else {
-            throw new IllegalArgumentException("graph must contain the start vertex");
+        this.graph = Objects.requireNonNull(graph);
+        this.weighted = weighted;
+        this.outEdgesTotalWeight = new HashMap<>();
+        this.hops = 0;
+        this.nextVertex = Objects.requireNonNull(vertex);
+        if (!graph.containsVertex(vertex)) {
+            throw new IllegalArgumentException("Random walk must start at a graph vertex");
         }
-
-        this.sinkReached = false;
-        this.random = Objects.requireNonNull(rng, "Random number generator cannot be null");
-    }
-
-    /**
-     * Check if this walk is exhausted. Calling {@link #next()} on exhausted iterator will throw
-     * {@link NoSuchElementException}.
-     * 
-     * @return <code>true</code>if this iterator is exhausted, <code>false</code> otherwise.
-     */
-    protected boolean isExhausted()
-    {
-        return maxSteps == 0;
-    }
-
-    /**
-     * Update data structures every time we see a vertex.
-     *
-     * @param vertex the vertex encountered
-     * @param edge the edge via which the vertex was encountered, or null if the vertex is a
-     *        starting point
-     */
-    protected void encounterVertex(V vertex, E edge)
-    {
-        maxSteps--;
+        this.maxHops = maxHops;
+        this.rng = rng;
     }
 
     @Override
     public boolean hasNext()
     {
-        return currentVertex != null && !isExhausted() && !sinkReached;
+        return nextVertex != null;
     }
 
     @Override
     public V next()
     {
-        if (!hasNext()) {
+        if (nextVertex == null) {
             throw new NoSuchElementException();
         }
-
-        Set<? extends E> potentialEdges = graph.outgoingEdgesOf(currentVertex);
-
-        // randomly select an edge from the set of potential edges.
-        E nextEdge = drawEdge(potentialEdges);
-        if (nextEdge != null) {
-            V nextVertex;
-            nextVertex = Graphs.getOppositeVertex(graph, nextEdge, currentVertex);
-            encounterVertex(nextVertex, nextEdge);
-            fireEdgeTraversed(createEdgeTraversalEvent(nextEdge));
-            fireVertexTraversed(createVertexTraversalEvent(nextVertex));
-            currentVertex = nextVertex;
-            return nextVertex;
-        } else {
-            sinkReached = true;
-            return currentVertex;
-        }
+        V value = nextVertex;
+        computeNext();
+        return value;
     }
 
-    /**
-     * Randomly draws an edges out of the provided set. In case of un-weighted walk, edge will be
-     * selected with uniform distribution across all outgoing edges. In case of a weighted walk,
-     * edge will be selected with probability respective to its weight across all outgoing edges.
-     * 
-     * @param edges the set to select the edge from
-     * @return the drawn edges or null if set is empty.
-     */
-    private E drawEdge(Set<? extends E> edges)
+    private void computeNext()
     {
-        if (edges.isEmpty()) {
-            return null;
+        if (hops >= maxHops) {
+            nextVertex = null;
+            return;
         }
 
-        int drawn;
-        List<E> list = new ArrayList<E>(edges);
-        if (isWeighted) {
-            Iterator<E> safeIter = list.iterator();
-            double border = random.nextDouble() * getTotalWeight(list);
-            double d = 0;
-            drawn = -1;
-            do {
-                d += graph.getEdgeWeight(safeIter.next());
-                drawn++;
-            } while (d < border);
+        hops++;
+        if (graph.outDegreeOf(nextVertex) == 0) {
+            nextVertex = null;
+            return;
+        }
+
+        E e = null;
+        if (weighted) {
+            double outEdgesWeight = outEdgesTotalWeight.computeIfAbsent(nextVertex, v -> {
+                return graph
+                    .outgoingEdgesOf(v).stream()
+                    .collect(Collectors.summingDouble(graph::getEdgeWeight));
+            });
+            double p = outEdgesWeight * rng.nextDouble();
+            double cumulativeP = 0d;
+            for (E curEdge : graph.outgoingEdgesOf(nextVertex)) {
+                cumulativeP += graph.getEdgeWeight(curEdge);
+                if (p <= cumulativeP) {
+                    e = curEdge;
+                    break;
+                }
+            }
         } else {
-            drawn = random.nextInt(list.size());
+            List<E> outEdges = new ArrayList<>(graph.outgoingEdgesOf(nextVertex));
+            e = outEdges.get(rng.nextInt(outEdges.size()));
         }
-        return list.get(drawn);
+        nextVertex = Graphs.getOppositeVertex(graph, e, nextVertex);
     }
 
-    private double getTotalWeight(Collection<E> edges)
-    {
-        double total = 0;
-        for (E e : edges) {
-            total += graph.getEdgeWeight(e);
-        }
-        return total;
-    }
 }
